Add carousel card and fixedvideowrapper
