+++ dotcom-rendering/src/web/layouts/FrontLayout.tsx
@@ -1,4 +1,5 @@
 import { css } from '@emotion/react';
+import { JSX } from '@emotion/react/jsx-runtime';
 import { ArticleDesign, ArticleDisplay, ArticlePillar } from '@guardian/libs';
 import {
 	brandBackground,
@@ -11,7 +12,11 @@ import { StraightLines } from '@guardian/source-react-components-development-kit
 import type { NavType } from '../../model/extract-nav';
 import type { DCRCollectionType, DCRFrontType } from '../../types/front';
 import { AdSlot } from '../components/AdSlot';
-import { FixedVideo } from '../components/FixedVideo';
+import {
+	CarouselCardMobile,
+	CarouselCardProps,
+} from '../components/CarouselCard';
+import { FixedVideoWrapper } from '../components/FixedVideoWrapper.importable';
 import { Footer } from '../components/Footer';
 import { FrontMostViewed } from '../components/FrontMostViewed';
 import { Header } from '../components/Header';
@@ -293,13 +298,21 @@ export const FrontLayout = ({ front, NAV }: Props) => {
 
 					if (collection.collectionType === 'fixed/video') {
 						return (
-							<FixedVideo
-								trails={trails}
-								key={ophanName}
-								containerName={collection.displayName}
-								ophanComponentLink={ophanComponentLink}
-								ophanComponentName={ophanName}
-							/>
+							<>
+								<Island deferUntil="visible">
+									<FixedVideoWrapper
+										trails={trails}
+										key={ophanName}
+										containerName={collection.displayName}
+										ophanComponentLink={ophanComponentLink}
+										ophanComponentName={ophanName}
+										CarouselCard={({
+											trail,
+											isFirst,
+										}: CarouselCardProps) => Element}
+									/>
+								</Island>
+							</>
 						);
 					}
